[fix] 질의응답 등록하기 API 구조 변경
