+++ src/pages/QnaCreatePage.tsx
@@ -1,8 +1,7 @@
 import { useState } from 'react'
 import { useNavigate } from 'react-router'
 
-import { getAccessToken } from '../api/api'
-import { createQnaPost } from '@/hooks/FetchQnaCreate'
+import { createQnaPost } from '@/api/qnaCreate'
 import CategorySelect from '../components/common/CategorySelect'
 import Editor from '../components/Editor/Editor'
 
@@ -41,13 +40,12 @@ export default function CommunityCreatePage() {
 
     try {
       setIsLoading(true)
-      const token = getAccessToken()
       const params = {
         title,
         content,
         category_id: detailCategoryId || subCategoryId || mainCategoryId,
       }
-      const data = await createQnaPost(params, token ?? '')
+      const data = await createQnaPost(params)
       console.log(data)
       alert('게시글이 등록되었습니다.')
       navigate(`/qnadetails/${data.question_id}`)
